// packages/sdk/src/generated/types/periodLimit.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export type PeriodLimit = {
  /** Window duration in seconds (0 = disabled) */
  durationSeconds: bigint;
  /** Maximum mint amount in this window */
  maxMintAmount: bigint;
  /** Maximum redeem amount in this window */
  maxRedeemAmount: bigint;
  /** Amount minted in current window */
  mintedAmount: bigint;
  /** Amount redeemed in current window */
  redeemedAmount: bigint;
  /** Window start timestamp */
  windowStart: bigint;
};

export type PeriodLimitArgs = {
  /** Window duration in seconds (0 = disabled) */
  durationSeconds: number | bigint;
  /** Maximum mint amount in this window */
  maxMintAmount: number | bigint;
  /** Maximum redeem amount in this window */
  maxRedeemAmount: number | bigint;
  /** Amount minted in current window */
  mintedAmount: number | bigint;
  /** Amount redeemed in current window */
  redeemedAmount: number | bigint;
  /** Window start timestamp */
  windowStart: number | bigint;
};

export function getPeriodLimitEncoder(): FixedSizeEncoder<PeriodLimitArgs> {
  return getStructEncoder([
    ["durationSeconds", getU64Encoder()],
    ["maxMintAmount", getU64Encoder()],
    ["maxRedeemAmount", getU64Encoder()],
    ["mintedAmount", getU64Encoder()],
    ["redeemedAmount", getU64Encoder()],
    ["windowStart", getI64Encoder()],
  ]);
}

export function getPeriodLimitDecoder(): FixedSizeDecoder<PeriodLimit> {
  return getStructDecoder([
    ["durationSeconds", getU64Decoder()],
    ["maxMintAmount", getU64Decoder()],
    ["maxRedeemAmount", getU64Decoder()],
    ["mintedAmount", getU64Decoder()],
    ["redeemedAmount", getU64Decoder()],
    ["windowStart", getI64Decoder()],
  ]);
}

export function getPeriodLimitCodec(): FixedSizeCodec<
  PeriodLimitArgs,
  PeriodLimit
> {
  return combineCodec(getPeriodLimitEncoder(), getPeriodLimitDecoder());
}